// Web/RMA-Web-Shell/projects/medicare/src/app/pmp-manager/models/pensioner-briefing-interview-form-detail.ts
export class PensionerInterviewFormDetail  {
    pensionerInterviewFormDetailsId: number;
    pensionerInterviewFormId: number;
    explainedCalculation: boolean;
    explainedPayDates: boolean;
    explainedProofOfLife: boolean;
    explainedIncreases: boolean;
    explainedMedicalTreatment: boolean;
    explainedPreAuthorisation: boolean;
    explainedMaintenance: boolean;
    suppliedBooklet: boolean;
    suppliedContactDetails: boolean;
    explainedChronicMedication: boolean;
    explainedTransportation: boolean;
    isActive: boolean;
    isDeleted: boolean;
    createdBy: string;
    createdDate: string;
    modifiedBy: string;
    modifiedDate: string;
}